Move district route
